requirements.txt
djangorestframework==3.16.0
django == 5.2.3
coreapi == 2.3.3
coreschema == 0.0.4
drf-yasg